+++ source/memory.h
@@ -17,7 +17,7 @@ private:
 	long long allocated;
 	size_t block_size;
 	long has_large_pages;
-	long long short_page_size, large_page_size;
+	long short_page_size, large_page_size;
 
 protected:
 	void *alloc(size_t *size);
